fix(server): coerce createdAt to JS number, narrow optionalEnum return type
- Cast `(EXTRACT(EPOCH FROM ...) * 1000)::float8` in auditLogs and
  notifications repos so pg-types parses the result as a JS number
  instead of returning numeric-as-string (matches the `createdAt: number`
  type and Fastify response schema).
- Make `optionalEnum` generic over `T extends string` so callers get
  the narrowed enum type back, removing the `as Language | null` cast
  in the settings PUT handler.

+++ server/repositories/auditLogsRepo.ts
@@ -47,7 +47,7 @@ export const list = async (
         al.entity_type as "entityType",
         al.entity_id as "entityId",
         al.ip_address as "ipAddress",
-        EXTRACT(EPOCH FROM al.created_at) * 1000 as "createdAt",
+        (EXTRACT(EPOCH FROM al.created_at) * 1000)::float8 as "createdAt",
         CASE WHEN jsonb_typeof(al.details) = 'object' THEN al.details ELSE NULL END as "details"
       FROM audit_logs al
       JOIN users u ON u.id = al.user_id${whereClause}
